ssr-app/pyproject.toml
[project]
name = "ssr-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
]
